==> R/scripts/01_processPgDetections.Rmd <==
---
title: "Process PamGuard Detections"
author: "Marie Zahn"
date: '2023-07-14'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

```{r load packages}
## load required packages
library(here)
library(ggplot2)
library(PAMpal)
library(dplyr)

#Set time zone to UTC
Sys.setenv(TZ = 'UTC')

# install PAMpal package from GitHub (or get latest version)
# devtools::install_github('TaikiSan21/PAMpal')
# # install BANTER package from GitHub
# devtools::install_github('ericarcher/banter')
```

Function that removes events that have no detections or very few (<30).

```{r function to filter out noise events, results = FALSE, include=FALSE}
filter_noise_acoustic_study <- function(Acoustic_Study){
  ## remove events that have no detections or very few (< 30 detections)
  events_list <- Acoustic_Study@events
  
  for (i in 1:length(events_list)) {
    # start counter at zero
    n_clicks <- 0
    
    # loop through detectors and get total number of clicks for each event
    for (j in 1:length(events_list[[i]]@detectors)) {
      if (length(events_list[[i]]@detectors) == 0) {
        n_clicks <- 0
      }
      if (length(events_list[[i]]@detectors) == 1) {
        n_clicks <- length(events_list[[i]]@detectors[[j]]$UID)
      }
      else if (length(events_list[[i]]@detectors) > 1){
        n_clicks_det <- length(events_list[[i]]@detectors[[j]]$UID)
        n_clicks <- n_clicks + n_clicks_det
      }
    }
    if (n_clicks<=30) {
      print(paste0("dropped event #", events_list[[i]]@id))
      Acoustic_Study@events[[events_list[[i]]@id]] <- NULL
    }
  }
  return(Acoustic_Study)
}

```

# Process Fisher and Kong Oscar subsets (for preliminary tests)

These subsets were run in PAMGuard with 4-pole 20 kHz high pass for trigger and 1 kHz pre-filter bandpass.
A 4-pole 20 kHz high pass filter is applied below for click parameter calculations.

```{r process Fisher and KO subsets}
#### Fisher Islands -------------
## version 2: with 1 kHz highpass filter, 20 kHz trigger, and SPL difference threshold
## version 3: with 1 kHz highpass filter, 20 kHz trigger
myPps <- PAMpalSettings(db = here("../../PAMGuard_ch4/subset_Fisher/v2/PAM20103_Zahn_narluga_Fisher-subset_20kHztrigger.sqlite3"),
                        binaries = here("../../PAMGuard_ch4/subset_Fisher/v2/Binaries"),
                        sr_hz = 144000, winLen_sec = 0.0025, filterfrom_khz = 20, filterto_khz = NULL)

Fisher19_subset_dets_20kHzPG_20kHzPP <- processPgDetections(myPps, mode ='time', 
                                              grouping = here('R/final_test/spreadsheets/Fisher_2019_species_ids_6hourly_subset.csv'), 
                                              format = '%m/%d/%Y %H:%M:%OS', id='2019Fisher')

## calculate ICI and add to object
Fisher19_subset_dets_20kHzPG_15kHzPP <- calculateICI(Fisher19_subset_dets_20kHzPG_15kHzPP, time = 'UTC', callType="click")

# save file
# save(Fisher_2019_data_0kHzPP, file='Rdata/Fisher_2019_0kHzPP_6hourly.rdata')
save(Fisher19_subset_dets_20kHzPG_15kHzPP, file='R/final_test/Rdata/Fisher19_subset_20kHzPG_15kHzPP_6hourly_v2.rdata')


#### Kong Oscar -------------
## version 2: with 1 kHz highpass filter, 20 kHz trigger, and SPL difference threshold
## version 3: with 1 kHz highpass filter, 20 kHz trigger

myPps <- PAMpalSettings(db = here("../../PAMGuard_ch4/subset_Kong/v3/PAM20103_Zahn_narluga_Kong-subset_20kHztrigger.sqlite3"),
                        binaries = here("../../PAMGuard_ch4/subset_Kong/v2/Binaries"),
                        sr_hz = 144000, winLen_sec = 0.0025, filterfrom_khz = 20,filterto_khz = NULL)

Kong19_subset_dets_20kHzPG_15kHzPP <- processPgDetections(myPps, mode ='time',
                                        grouping = here('R/final_test/spreadsheets/KO_2019_species_ids_6hourly_subset.csv'),
                                        format = '%m/%d/%Y %H:%M:%OS', id='2019KongOscar')

## calculate ICI and add to object
Kong19_subset_dets_20kHzPG_15kHzPP <- calculateICI(Kong19_subset_dets_20kHzPG_15kHzPP, time = 'UTC', callType="click")

# save file
save(Kong19_subset_dets_20kHzPG_15kHzPP, file=here('R/final_test/Rdata/KO19_subset_20kHzPG_15kHzPP_6hourly_v2.rdata'))
```

# Process full datasets 1 hr

This is processing datasets used with the latest version of PAMGuard 2.02.09

These data include only clicks from Detectors 2 and 3.
Data were processed with a 4-pole 1 kHz pre-filter high pass and 20 kHz trigger highpass filter.
Detector 2 clicks were only classified if there was >5 dB difference between 10-20 and 30-40 kHz bands.

```{r process Fisher and Kong Oscar data}
#### Fisher Islands -------------
myPps <- PAMpalSettings(db = here("../Fisher/PAM20209_Zahn_narluga_fisher.sqlite3"),
                        binaries = here("../Fisher/Binaries"),
                        sr_hz = 144000, winLen_sec = 0.0025, filterfrom_khz = 0, filterto_khz = NULL)

Fish_dets_0kHzPP <- processPgDetections(myPps, mode ='time',
                                         grouping = here('R/spreadsheets/1hr/Fisher_2019_species_ids_1hr_wavTime_labeled.csv'),
                                         format = '%m/%d/%Y %H:%M:%OS', id='2019Fisher')

## calculate ICI and add to object
fish_dets_0kHzPP <- calculateICI(Fish_dets_0kHzPP, time = 'UTC', callType="click")

# save file
save(fish_dets_0kHzPP, file=here('R/Rdata/1hr/Fish_20kHzPG_0kHzPP_1hr_wavTime_unfiltered_labeled.rdata'))

# filter out events that have no detections and those that have less than 30 detections
fish_dets_0kHzPP_filter <- filter_noise_acoustic_study(fish_dets_0kHzPP)

# save file
save(fish_dets_0kHzPP_filter, file=here('R/Rdata/1hr/Fish_20kHzPG_0kHzPP_1hr_wavTime_filtered_labeled.rdata'))

#### Kong Oscar -------------
myPps <- PAMpalSettings(db = here("../Kong_Oscar/PAM20209_Zahn_narluga_kong.sqlite3"),
                        binaries = here("../Kong_Oscar/Binaries"),
                        sr_hz = 144000, winLen_sec = 0.0025, filterfrom_khz = 0,filterto_khz = NULL)

Kong_dets_20kHzPP <- processPgDetections(myPps, mode ='time',
                                         grouping = here('R/spreadsheets/1hr/Kong_2019_species_ids_1hr_wavTime_labeled.csv'),
                                         format = '%m/%d/%Y %H:%M:%OS', id='2019KongOscar')

## calculate ICI and add to object
kong_dets_20kHzPP <- calculateICI(Kong_dets_20kHzPP, time = 'UTC', callType="click")

# save file
save(kong_dets_20kHzPP, file=here('R/Rdata/1hr/Kong_20kHzPG_0kHzPP_1hr_wavTime_unfiltered_labeled.rdata'))

# filter out events that have no detections and those that have less than 30 detections
kong_dets_20kHzPP_filter <- filter_noise_acoustic_study(kong_dets_20kHzPP)

# save file
save(kong_dets_20kHzPP_filter, file=here('R/Rdata/1hr/Kong_20kHzPG_0kHzPP_1hr_wavTime_filtered_labeled.rdata'))

```


```{r check # events for each group}
## Fisher
fish_beluga_dets <- Fisher19_dets_20kHzPG_15kHzPP %>% filter(species=='X045')
fish_beluga_dets_filter <- Fish_filtered %>% filter(species=='X045')
fish_narwhal_dets <- Fisher19_dets_20kHzPG_15kHzPP %>% filter(species=='X085')
fish_narwhal_dets_filter <- Fish_filtered %>% filter(species=='X085')
fish_noise_dets <- Fish_filtered %>% filter(species=='NOISE')

## filtered
length(fish_beluga_dets_filter@events)
length(fish_narwhal_dets_filter@events)

## unfiltered
length(fish_beluga_dets@events)
length(fish_narwhal_dets@events)

## Kong Oscar
kong_beluga_dets <- Kong19_dets_20kHzPG_15kHzPP %>% filter(species=='X045')
kong_beluga_dets_filter <- Kong_filtered %>% filter(species=='X045')
kong_narwhal_dets <- Kong19_dets_20kHzPG_15kHzPP %>% filter(species=='X085')
kong_narwhal_dets_filter <- Kong_filtered %>% filter(species=='X085')
kong_noise_dets <- Kong_filtered %>% filter(species=='NOISE')

## filtered
length(kong_beluga_dets_filter@events)
length(kong_narwhal_dets_filter@events)

## unfiltered
length(kong_beluga_dets@events) # this should be =5, but event 252 had no detections so it is =4
length(kong_narwhal_dets@events)
```
